Main Menu Script
Created a main menu script that just polls user input. (AS Player one, KL Player two), and tells the users that each player is ready when both keys are held down. An event manager should be made to tell the game that the game should begin when both players are ready.